requirements.txt
streamlit 
langchain 
google-generativeai 
PyPDF2 
python-dotenv
